requirements.txt
pandas
python-whois
tldextract
matplotlib
seaborn
scikit-learn
category-encoders
fastapi
uvicorn[standard]
joblib
numpy
requests
python-multipart
python-dotenv